refactor(simplifier): extract rime::Opencc into opencc.h / opencc.cc
rime::Opencc was a private class defined inside simplifier.cc, making
it impossible to test in isolation without spinning up a full RIME
Engine. Move it to its own header and source file so unit and
integration tests can instantiate it directly.

- src/rime/gear/opencc.h: class declaration; forward-declares
  opencc::Converter and opencc::Dict to keep OpenCC headers out of the
  public interface (preserving the VS2015 constraint noted in the
  original code)
- src/rime/gear/opencc.cc: full implementation with all OpenCC includes
- src/rime/gear/simplifier.cc: replace the inline class with
  #include <rime/gear/opencc.h>; retain #include <opencc/Exception.hpp>
  for the catch block in SimplifierComponent::Create()
